Split `lazy` into `hydration` and `lazy`
